+++ lib/error.ml
@@ -14,9 +14,31 @@ type t = F.formatter -> unit
 
 let pp = ( |> )
 
-exception Acutis_error of t
+exception Fatal of t
 
-let raise x = raise @@ Acutis_error x
+type _ Effect.t += Warn : t -> unit Effect.t
+
+module ES = Effect.Shallow
+
+type 'a res = t Seq.t * 'a option
+
+let rec continue : type a. t Seq.t -> (unit, a) ES.continuation -> a res =
+ fun warnings k ->
+  let retc x = (warnings, Some x) in
+  let exnc = function
+    | Fatal e -> (Seq.append warnings @@ Seq.return e, None)
+    | e -> raise e
+  in
+  let effc : type b. b Effect.t -> ((b, a) ES.continuation -> a res) option =
+    function
+    | Warn w -> Some (continue (Seq.append warnings @@ Seq.return w))
+    | _ -> None
+  in
+  ES.continue_with k () { retc; exnc; effc }
+
+let handle f x = continue Seq.empty (ES.fiber (fun () -> f x))
+let fatal x = raise @@ Fatal x
+let warn x = Effect.perform @@ Warn x
 
 let msg ~kind loc t =
   F.dprintf "@[<v>File \"%s\", %a@;%s.@;%t@]" (Loc.fname loc) Loc.pp loc kind t
@@ -28,19 +50,19 @@ let msg_syntax = msg ~kind:"Syntax error"
 
 let lex_unexpected lexbuf c =
   let f = F.dprintf "Unexpected character: %C" c in
-  raise @@ msg_syntax (loc_of_lexbuf lexbuf) f
+  fatal @@ msg_syntax (loc_of_lexbuf lexbuf) f
 
 let lex_bad_int lexbuf s =
   let f = F.dprintf "Invalid integer: %s" s in
-  raise @@ msg_syntax (loc_of_lexbuf lexbuf) f
+  fatal @@ msg_syntax (loc_of_lexbuf lexbuf) f
 
 let lex_unterminated_comment lexbuf =
   let f = F.dprintf "@[%a@]" F.pp_print_text "Unterminated comment." in
-  raise @@ msg_syntax (loc_of_lexbuf lexbuf) f
+  fatal @@ msg_syntax (loc_of_lexbuf lexbuf) f
 
 let lex_unterminated_string lexbuf =
   let f = F.dprintf "@[%a@]" F.pp_print_text "Unterminated string." in
-  raise @@ msg_syntax (loc_of_lexbuf lexbuf) f
+  fatal @@ msg_syntax (loc_of_lexbuf lexbuf) f
 
 let parse_error i loc =
   let f =
@@ -49,60 +71,61 @@ let parse_error i loc =
       F.dprintf "@[%a@]" F.pp_print_text (String.trim mess)
     with Not_found -> F.dprintf "Unexpected token."
   in
-  raise @@ msg ~kind:"Parse error" loc f
+  fatal @@ msg ~kind:"Parse error" loc f
 
 let msg_ty = msg ~kind:"Type error"
+let msg_ty_warn = msg ~kind:"Type warning"
 
 let dup_record_key loc key =
   let f = F.dprintf "Duplicate field '%a'." Pp.field key in
-  raise @@ msg_ty loc f
+  fatal @@ msg_ty loc f
 
 let extra_record_tag =
   let f = F.dprintf "This tagged record has multiple tags." in
-  fun loc -> raise @@ msg_ty loc f
+  fun loc -> fatal @@ msg_ty loc f
 
 let mismatch pp a b t =
   F.dprintf "Type mismatch.@;Expected:@;<1 2>%a@;Received:@;<1 2>%a" pp a pp b t
 
-let type_mismatch loc pp a b = raise @@ msg_ty loc (mismatch pp a b)
+let type_mismatch loc pp a b = fatal @@ msg_ty loc (mismatch pp a b)
 
 let bad_block loc =
   let f =
     F.dprintf "Template blocks are not allowed in a destructure pattern."
   in
-  raise @@ msg_ty loc f
+  fatal @@ msg_ty loc f
 
 let bad_field loc =
   let f =
     F.dprintf "Record '.' access is not allowed in a destructure pattern."
   in
-  raise @@ msg_ty loc f
+  fatal @@ msg_ty loc f
 
 let missing_field loc key pp ty =
   let f =
     F.dprintf "This is missing key '%a' of type:@;<1 2>%a" Pp.field key pp ty
   in
-  raise @@ msg_ty loc f
+  fatal @@ msg_ty loc f
 
 let underscore_in_construct =
   let f = F.dprintf "Underscore ('_') is not a valid name." in
-  fun loc -> raise @@ msg_ty loc f
+  fun loc -> fatal @@ msg_ty loc f
 
 let name_bound_too_many loc s =
   let f = F.dprintf "The name '%s' is already bound in this pattern." s in
-  raise @@ msg_ty loc f
+  fatal @@ msg_ty loc f
 
 let var_missing loc v =
   let f = F.dprintf "Variable '%s' must occur in each 'with' pattern." v in
-  raise @@ msg_ty loc f
+  fatal @@ msg_ty loc f
 
 let var_unused loc s =
   let f = F.dprintf "This variable is bound but never used:@;<1 2>%s" s in
-  raise @@ msg_ty loc f
+  warn @@ msg_ty_warn loc f
 
 let pat_num_mismatch =
   let f = F.dprintf "Pattern count mismatch." in
-  fun loc -> raise @@ msg_ty loc f
+  fun loc -> fatal @@ msg_ty loc f
 
 let map_pat_num_mismatch =
   let f =
@@ -110,48 +133,48 @@ let map_pat_num_mismatch =
       "@[Expressions 'map' and 'map_dict' can only have one or two@ patterns@ \
        for@ each@ 'with'@ expression.@]"
   in
-  fun loc -> raise @@ msg_ty loc f
+  fun loc -> fatal @@ msg_ty loc f
 
 let component_name_mismatch loc a b =
   let f =
     F.dprintf
       "Component name mismatch.@;Expected:@;<1 2>%s.@;Received:@;<1 2>%s." a b
   in
-  raise @@ msg_ty loc f
+  fatal @@ msg_ty loc f
 
 let component_extra_prop loc name s =
   let f =
     F.dprintf "Component '%s' does not accept this prop:@;<1 2>%a." name
       Pp.field s
   in
-  raise @@ msg_ty loc f
+  warn @@ msg_ty_warn loc f
 
 let interface_duplicate loc id =
   let f = F.dprintf "Prop '%s' is already defined in the interface." id in
-  raise @@ msg_ty loc f
+  fatal @@ msg_ty loc f
 
 let interface_bad_name loc id =
   let f = F.dprintf "There is no type named '%s'." id in
-  raise @@ msg_ty loc f
+  fatal @@ msg_ty loc f
 
 let interface_untagged_union loc =
   let f = F.dprintf "You cannot union records without a '@' tag field." in
-  raise @@ msg_ty loc f
+  fatal @@ msg_ty loc f
 
 let interface_unmatched_tags loc s1 s2 =
   let f =
     F.dprintf "This record has tag field '@%a' instead of '@%a'." Pp.field s2
       Pp.field s1
   in
-  raise @@ msg_ty loc f
+  fatal @@ msg_ty loc f
 
 let interface_duplicate_tag loc pp tag =
   let f = F.dprintf "Tag value '%a' is already used in this union." pp tag in
-  raise @@ msg_ty loc f
+  fatal @@ msg_ty loc f
 
 let interface_open_bool_union loc =
   let f = F.dprintf "Unions with boolean tags cannot be opened with '...'." in
-  raise @@ msg_ty loc f
+  fatal @@ msg_ty loc f
 
 let interface_type_mismatch loc k pp a b =
   let f =
@@ -165,7 +188,7 @@ let interface_type_mismatch loc k pp a b =
        <1 2>%a"
       k pp a pp b
   in
-  raise @@ msg_ty loc f
+  fatal @@ msg_ty loc f
 
 let interface_missing_prop loc k pp ty =
   let f =
@@ -177,13 +200,14 @@ let interface_missing_prop loc k pp ty =
        <1 2>%a"
       k pp ty
   in
-  raise @@ msg_ty loc f
+  fatal @@ msg_ty loc f
 
 let msg_match = msg ~kind:"Matching error"
+let msg_match_warn = msg ~kind:"Matching warning"
 
 let unused_case loc =
   let f = F.dprintf "This match case is unused." in
-  raise @@ msg_match loc f
+  warn @@ msg_match_warn loc f
 
 let parmatch loc pp_pat pat =
   let f =
@@ -193,7 +217,7 @@ let parmatch loc pp_pat pat =
        <1 2>%a"
       pp_pat pat
   in
-  raise @@ msg_match loc f
+  fatal @@ msg_match loc f
 
 let duplicate_name s =
   let f =
@@ -201,7 +225,7 @@ let duplicate_name s =
       "@[<v>Compile error.@;There are multiple components with the name '%s'.@]"
       s
   in
-  raise @@ f
+  fatal @@ f
 
 let msg_compile = F.dprintf "@[<v>Compile error.@;%t@]"
 let pp_sep ppf () = F.fprintf ppf "@ -> "
@@ -212,22 +236,22 @@ let cycle stack =
       (F.pp_print_list ~pp_sep F.pp_print_string)
       (List.rev stack)
   in
-  Acutis_error (msg_compile f)
+  Fatal (msg_compile f)
 
 let missing_component stack name =
   let f =
     F.dprintf "Missing template:@;<1 2>%s@;Required by:@;<1 2>%s" name
       (List.hd stack)
   in
-  Acutis_error (msg_compile f)
+  Fatal (msg_compile f)
 
 let internal ~__POS__:(file, lnum, cnum, enum) =
   F.kdprintf @@ fun t ->
-  raise
+  fatal
   @@ F.dprintf
        "This is a bug in the compiler. Please contact the Acutis developer.@;\
         @[OCaml source file %S, line %d, characters %d-%d.@]@;\
         @[%t@]"
        file lnum cnum enum t
 
-let raise_fmt x = F.kdprintf raise x
+let raise_fmt x = F.kdprintf fatal x