+++ test/printjs/printjs.t/run.t
@@ -1812,9 +1812,9 @@
             )
           );
       buf$0.contents += "\n";
-      return (buf$0.contents);
+      return (Promise.resolve(buf$0.contents));
     } else {
-      throw (new Error(errors$0.contents));
+      return (Promise.reject(new Error(errors$0.contents)));
     }
   };